Fixed order view bug
Fixed a bug where you would not see all pizzas which were ordered in the Order View.
Fixed a bug where there would be some overlapping mistake in the Shopping Cart View.

+++ DeliverApp/PizzaAppDelivery/Data/DownloadOrders.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct SinglePizzaOrdered: Codable, Hashable {
     var pizzaId: Int32
-    var sizeIndex: Int16
+    var sizeIndex: Int8
 }
 
 struct SingleOrder: Codable, Hashable {
